Added Redis-backed caching
Added transparent caching support to the client. Currently only
supports a Redis backend. Expiry time can be configured, and cache can
be bypassed (cache_bypass=True) or completely skipped (cache_skip=True)
on any standard client method.

A Session must now be passed as an argument if default options need to
be changed at init. Also a timeout is highly recommended to avoid
locking issues. The session timeout should be set to half of what the
cache lock timeout is.

Requires the 'redis' and 'msgpack' modules.